Main2

+++ admin/views/cart/index.php
@@ -13,20 +13,19 @@
         <div class="card-header py-3">
             <h6 class="m-0 font-weight-bold text-primary">DataTables</h6>
         </div>
+        <?php if (isset($_SESSION['success'])) : ?>
+
+
+            <div class="alert alert-success">
+                <?= $_SESSION['success'] ?>
+            </div>
+
+            <?php unset($_SESSION['success']) ?>
+
+        <?php endif; ?>
         <div class="card-body">
-            <?php if (isset($_SESSION['success'])) : ?>
-
-
-                <div class="alert alert-success">
-                    <?= $_SESSION['success'] ?>
-                </div>
-
-                <?php unset($_SESSION['success']) ?>
-
-            <?php endif; ?>
             <div class="table-responsive">
                 <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
-
                     <thead>
                         <tr>
                             <th>ID</th>
@@ -35,8 +34,6 @@
                             <th>Số điện thoại</th>
                             <th>Địa chỉ</th>
                             <th>Tổng tiền hàng</th>
-                            <th>Ngày tạo đơn</th>
-                            <th>Ngày cập nhật cùng</th>
                             <th>Shipping</th>
                             <th>Trạng thái đơn hàng</th>
                             <th>Trạng thái thanh toán</th>
@@ -44,11 +41,8 @@
                             <th>Action</th>
                         </tr>
                     </thead>
-
-                    <?php foreach ($carts as $cart) : ?>
-
-                        <tbody>
-
+                    <tbody>
+                        <?php foreach ($carts as $cart) : ?>
                             <tr>
                                 <td><?= $cart['id'] ?></td>
                                 <td><?= $cart['user_name'] ?></td>
@@ -56,8 +50,6 @@
                                 <td><?= $cart['user_phone'] ?></td>
                                 <td><?= $cart['user_address'] ?></td>
                                 <td><?= number_format($cart['total_bill'])  ?>VND</td>
-                                <td><?= $cart['created_at'] ?></td>
-                                <td><?= $cart['updated_at'] ?></td>
                                 <td><?= $cart['shipping'] ?></td>
                                 <td class="col-sm-2">
                                     <span class=" text-info">
@@ -66,22 +58,23 @@
                                 </td>
 
 
-                                <td> <?= ($cart['status_payment']  == 0) ? '<span class="badge badge-warning">Chưa thanh toán</span>' : ($cart['status_payment'] == 1 ? '<span class="badge badge-success">Đã thanh toán</span>' : ($cart['status_payment'] == -1 ? '<span class="badge badge-danger">Đơn hàng đã hủy</span>' : '')) ?></td>
+                                <td> <?= ($cart['thanhtoan']  == 0) ? '<span class="badge badge-warning">Chưa thanh toán</span>' : ($cart['thanhtoan'] == 1 ? '<span class="badge badge-success">Đã thanh toán</span>' : ($cart['status_delivery'] == -1 ? '<span class="badge badge-danger">Đơn hàng đã hủy</span>' : '')) ?></td>
 
 
 
                                 <td>
                                     <a class="btn btn-info mt-1" href="<?= BASE_URL_ADMIN ?>?act=cart-detail&id=<?= $cart['id'] ?>">Xem chi tiết</a>
-                                    <a class="btn btn-warning mt-1" href="<?= BASE_URL_ADMIN ?>?act=cart-update&id=<?= $cart['id'] ?>">Cập nhật</a>
-                                    <a class="btn btn-danger mt-1" href="<?= BASE_URL_ADMIN ?>?act=cart-delete&id=<?= $cart['id'] ?>" onclick="return confirm('Bạn có chắc chắn muốn xóa không?')">Hủy bỏ đơn hàng</a>
+                                    <?php if ($cart['status_delivery'] != -1 && $cart['status_delivery'] != 3) { ?>
+                                        <a class="btn btn-warning mt-1" href="<?= BASE_URL_ADMIN ?>?act=cart-update&id=<?= $cart['id'] ?>">Cập nhật</a>
+                                    <?php } ?>
+                                    <?php if ($cart['status_delivery'] == 0) { ?>
+                                        <a class="btn btn-danger mt-1" href="<?= BASE_URL_ADMIN ?>?act=cart-delete&id=<?= $cart['id'] ?>" onclick="return confirm('Bạn có chắc chắn muốn xóa không?')">Hủy bỏ đơn hàng</a>
+                                    <?php } ?>
                                 </td>
 
                             </tr>
-
-                        </tbody>
-
-                    <?php endforeach; ?>
-
+                        <?php endforeach; ?>
+                    </tbody>
                 </table>
             </div>
         </div>
